- Move schema manager to its own file - rename tst parser to tdr parser
